+++ src/main/java/com/GAB1NMACHINE/microwaves/entity/User.java
@@ -1,10 +1,14 @@
-package com.GAB1NMACHINE.MicoWave.entity;
+package com.GAB1NMACHINE.microwaves.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import lombok.Getter;
+import lombok.Setter;
 
+@Getter
+@Setter
 @Entity
 public class User {
     @Id
@@ -19,10 +23,4 @@ public class User {
 
     private String hashedPassword;
 
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String johnDoe) {
-    }
 }
