+++ src/main/java/com/GAB1NMACHINE/microwaves/repository/MicroWaveRepository.java
@@ -1,6 +1,6 @@
-package repository;
+package com.GAB1NMACHINE.microwaves.repository;
 
-import com.GAB1NMACHINE.MicoWave.entity.MicroWave;
+import com.GAB1NMACHINE.microwaves.entity.MicroWave;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
