+++ src/main/java/com/GAB1NMACHINE/microwaves/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package repository;
+package com.GAB1NMACHINE.microwaves.repository;
 
-import com.GAB1NMACHINE.MicoWave.entity.User;
+import com.GAB1NMACHINE.microwaves.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
